Delete commentd not workigng

+++ HappyNews/src/main/java/com/fdmgroup/HappyNews/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.fdmgroup.HappyNews.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
@@ -23,10 +24,22 @@ public class CommentService {
 		
 	}
 	
-	public void deleteComment (int commentId) {
-		commentRepository.deleteById(commentId);
+	/*public void deleteComment (int commentId) {
+		System.out.println("we are in the service====================================================" + commentId);
+		Optional<Comment> optinal =commentRepository.findById(commentId);
+		if(optinal.isPresent()) {
+			List<Comment> listOfChildComments = commentRepository.findByParentComment(optinal.get());
+			if(!listOfChildComments.isEmpty() || listOfChildComments.size()>0) {
+				for(Comment comment: listOfChildComments) {
+					commentRepository.delete(comment);
+				}
+			}
+			
+			commentRepository.deleteById(commentId);;
+		}
 		
-	}
+		
+	}*/
 	
 	public List <Comment> listOfCommentsForArticle(Article article){
 		return commentRepository.findByArticle(article);
